(Config) Move config manager out of application module
