+++ scripts/start_therm.py
@@ -4,7 +4,8 @@ import json
 
 from w1thermsensor import W1ThermSensor
 
-from sensormesh.application import Controller, ConfigManager
+from sensormesh.config import ConfigManager
+from sensormesh.application import Controller
 from sensormesh.sources import DataSourceWrapper
 from sensormesh.console import ConsoleDisplay
 from sensormesh.thingspeak import ThingSpeakLogger
